combined buildArticleHeader into ArticleMeta block for more consistent social icons
