+++ blocks/article-meta/article-meta.css
@@ -1,3 +1,33 @@
+.as-article-header {
+  margin-top: 128px;
+}
+
+/* title */
+.article-title-section {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-xs);
+}
+
+.article-title-section > a {
+  font-weight: var(--type-detail-all-weight);
+  color: var(--color-gray-600);
+  font-size: var(--type-detail-m-size);
+  line-height: var(--type-detail-m-lh);
+  text-transform: uppercase;
+  text-decoration: none;
+  letter-spacing: .1em;
+  width: fit-content;
+}
+
+.article-title-section h1 {
+  margin: 0;
+  color: var(--color-gray-800);
+  font-size: var(--type-heading-m-size);
+  line-height: var(--type-heading-m-lh);
+}
+
+/* author info */
 .article-byline {
   display: flex;
   flex-wrap: wrap;
@@ -119,7 +149,47 @@
   color: var(--color-info-accent-hover);
 }
 
+/* image */
+.article-feature-image {
+  padding: 0;
+  margin: auto;
+  width: 100%;
+  max-width: 1440px;
+}
+
+.article-feature-image img {
+  width: 100%;
+  max-height: 580px;
+  object-fit: cover;
+}
+
+.article-feature-image .caption {
+  box-sizing: border-box;
+  display: inline-block;
+  width: 100%;
+  text-align: center;
+  padding: 0 var(--spacing-m);
+}
+
+.as-article-header .article-title-section,
+.as-article-header .article-byline {
+  max-width: min(var(--grid-container-width), 800px);
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.as-article-header .share .icon-container {
+  padding: 0 0 var(--spacing-m);
+  border: 0;
+  gap: var(--spacing-s);
+}
+
 @media (min-width: 600px) {
+  .article-title-section h1 {
+    font-size: var(--type-heading-xl-size);
+    line-height: var(--type-heading-xl-lh);
+  }
+
   .article-byline {
     margin: var(--spacing-m) auto;
     padding-bottom: var(--spacing-xs);
@@ -128,6 +198,11 @@
 
   .article-byline .article-byline-info {
     flex: 1 1 auto;
+  }
+
+  .article-meta.as-article-header .article-byline {
+    padding: var(--spacing-m) 0 var(--spacing-l);
+    margin: 0 auto;
   }
 
   .article-meta .share {
@@ -139,6 +214,11 @@
     margin-top: 0;
     gap: var(--spacing-s);
     border: 0;  
+  }
+
+  .article-meta.as-article-header .share .icon-container { 
+    padding: 0;
+    gap: var(--spacing-xs);
   }
 }
 
@@ -168,5 +248,15 @@
     justify-content: flex-start;
     gap: var(--spacing-xs);
     padding: var(--spacing-xxs) 0 0;
+  }
+
+  .article-meta.as-article-header .share {
+    width: fit-content;
+  }
+}
+
+@media (min-width: 1440px) {
+  .article-feature-image .figure-feature img {
+    object-fit: contain;
   }
 }
